requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Templates
jinja2>=3.1.0

# HTTP Client (for external API calls)
httpx>=0.26.0

# Google Gemini API
google-genai>=1.0.0
Pillow>=10.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
